==> taxi-trips/02-web_scraping.ipynb <==
from bs4 import BeautifulSoup
import pandas as pd
import requests

web_source = 'https://en.wikipedia.org/wiki/Community_areas_in_Chicago'

response = requests.get(url=web_source)
soup = BeautifulSoup(response.content, "html.parser")


#Oldal cím kiirása
title = soup.find(id="firstHeading")


title
print(title.string)

print(soup.prettify())

#Táblázatok megkeresése
soup.find_all('table')

#Táblázatok osztálynevei

# <table class="wikitable sortable plainrowheaders mw-datatable" style="text-align:right">
           
for table in soup.find_all('table'):
    print(table.get('class'))

table = soup.find("table", class_="wikitable sortable plainrowheaders mw-datatable")
print(table.prettify())

#A tábla első két sora és az utoldó sora nem érdekel bennünket, ezért a [2:-1] szeletelés
for row in table.tbody.find_all("tr")[2:-1]:
    print(row)
    print('-' * 100)

for row in table.tbody.find_all("tr")[2:-1]:
    cells = row.find_all("td")
    print(cells[0].get_text(strip = True))

for row in table.tbody.find_all("tr")[2:-1]:
    header_cells = row.find("th")
    print(header_cells.a.get_text(strip=True))

data = []
for row in table.tbody.find_all("tr")[2:-1]:
    cells = row.find_all("td")
    header_cells = row.find("th")
    area_code = cells[0].get_text(strip = True)
    area_name = header_cells.a.get_text(strip=True)
    print(area_code)
    print(area_name)
    
    data.append({'Area code': area_code, 'Area name': area_name})

data

community_areas = pd.DataFrame(data)
community_areas['Area code'] = community_areas['Area code'].astype('int')
community_areas

community_areas.to_csv(r'..\data\community_areas_master.csv', index=False)
