==> taxi-trips/05-_date_dim_table.ipynb <==
import pandas as pd
from datetime import datetime

#Calendar table készítése
start_date = datetime(2023,1,1)
end_date = datetime(2028,1,1)

date_range = pd.date_range(start_date, end_date)
date_df = pd.DataFrame(date_range, columns=['date'])
date_df

holidays = [
    "11",
    "1012",
    "74",
    "1111",
    "911",
    "1225",
    "222",
    "214",
    "317",
    "1031"
]

date_df['year'] = date_df['date'].dt.year
date_df['month'] = date_df['date'].dt.month
date_df['month'] = date_df['month'].astype('str')
date_df['month name'] = date_df['date'].dt.month_name()
date_df['day'] = date_df['date'].dt.day
date_df['day'] = date_df['day'].astype('str')
date_df['day of week'] = date_df['date'].dt.day_of_week + 1
date_df['day of name'] = date_df['date'].dt.day_name()
date_df['is_weekend'] = date_df['day of week'].isin([6,7])
date_df['is_holiday_helper'] = date_df['month'] + date_df['day']
date_df['is_holiday'] = date_df['is_holiday_helper'].isin(holidays)
date_df


date_df.drop(['is_holiday_helper'],axis=1, inplace=True)
date_df.to_csv('../data/date_dim.csv', index=False)
